Add IT for caregiver profile page
