+++ packages/yourloops/test/integration/assert/profile.ts
@@ -25,8 +25,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { screen, within } from '@testing-library/react'
+import { act, screen, within } from '@testing-library/react'
 import { CountryCodes } from '../../../lib/auth/models/country.model'
+import userEvent from '@testing-library/user-event'
+import { AuthApi } from '../../../lib/auth/auth.api'
 
 interface CommonFieldsHtmlElements {
   firstNameInput: HTMLElement
@@ -115,4 +117,34 @@ export const checkPatientProfilePage = (country: CountryCodes): PatientFieldsHtm
   expect(ssnInput).not.toBeInTheDocument()
 
   return inputs
+}
+
+export const checkPasswordChangeRequest = async (): Promise<void> => {
+  const changePasswordCategoryTitle = screen.getByText('Security')
+  expect(changePasswordCategoryTitle).toBeVisible()
+
+  const changePasswordInfoLabel = screen.getByText('By clicking this button, you will receive an e-mail allowing you to change your password.')
+  expect(changePasswordInfoLabel).toBeVisible()
+
+  const changePasswordButton = screen.getByRole('button', { name: 'Change password' })
+  expect(changePasswordButton).toBeVisible()
+
+  await act(async () => {
+    await userEvent.click(changePasswordButton)
+  })
+
+  const changePasswordEmailSuccessfulSnackbar = screen.getByTestId('alert-snackbar')
+  expect(changePasswordEmailSuccessfulSnackbar).toHaveTextContent('E-mail sent successfully')
+
+  const changePasswordEmailSuccessfulSnackbarCloseButton = within(changePasswordEmailSuccessfulSnackbar).getByTitle('Close')
+
+  await userEvent.click(changePasswordEmailSuccessfulSnackbarCloseButton)
+
+  jest.spyOn(AuthApi, 'sendResetPasswordEmail').mockRejectedValueOnce('Error')
+  await act(async () => {
+    await userEvent.click(changePasswordButton)
+  })
+
+  const changePasswordEmailFailedSnackbar = screen.getByTestId('alert-snackbar')
+  expect(changePasswordEmailFailedSnackbar).toHaveTextContent('Impossible to send the change password e-mail. Please try again later.')
 }
